add validations and mostly working sync

+++ src/main/java/pl/edu/pg/eti/kask/kzawora/housing_community/resource/HousingCommunityResource.java
@@ -65,7 +65,7 @@ public class HousingCommunityResource {
                         .build())
         );
 
-        int size = service.countHousingCommunities();
+        int size = (int) service.countHousingCommunities();
 
         EmbeddedResource.EmbeddedResourceBuilder<List<HousingCommunity>> builder = EmbeddedResource.<List<HousingCommunity>>builder()
                 .embedded("housingCommunities", housingCommunities);
